Split user tests into another module
