--- tests/api/v1/test_permissions.py
@@ -1,7 +1,0 @@
-from uuid import uuid4
-from json import dumps
-
-
-# def test_get_permissions(client):
-#     user_id = uuid4()
-#     workspace_id = uuid4()
